Initial commit and release of `assay`
This commit is the first release of `assay` which is designed to solve a
lot of frustrations I had around testing in Rust. Mainly that:

- Tests all run in the same process which means setting env vars or
  changing the working dir affects all of the tests meaning you have to
  resort to things like `cargo test -- --test-threads=1` or using some
  kind of mutex whereby you lose the parallelization of running the test
  suite
- Setting up a temporary file system to run things in for a test and
  having the test run inside it is a pain to setup and being relative to
  it by using `std::env::set_working_dir` is prone to the above issues
- Including fixtures in a test, let alone multiple, can get a bit verbose
- Setting up and tearing down the same thing for each test can be a lot
- Want to run `async` tests? There's no runtime and you have to do
  setup just to run it.
- Using `?` in your test means putting `-> Result<(), Box<dyn std::error::Error>>`
  on every test and it can be tedious
- `assert_eq`/`assert_ne` output can be hard to grok and see why
  something is equal/not equal

`assay` is a proc macro that solves all of these problems by getting rid
of the repetitive nature of it, allowing setup and teardown, per test
env vars, and more. While there's lots that can be done to further
improve `assay` it's in a great enough spot to be used today.