Remove rusty_fork in favor of handrolled solution
This commit changes up how the assay macro works by forking the code
into it's Command process instead of using rusty_fork which did the same
thing under the hood. rusty_fork is both outdated and not the most
ergonomic to work with. With these changes our handrolled solution is a
bit easier to work with as a library author and makes it easier for us
to make sure tests fail the way we expect and capture that output
properly. In order to support this feature we also:

- Added support for the ignore attribute in assay
- Added a python script to run the main test suite and two tests we
  expect to fail so that we know the Command forking actually works
- Upgraded the dependencies to more modern versions since it has been
  2 years since the last commit

Closes #12
Closes #3