Fix panics and pretty_assert not outputting stdout
This commit fixes assay so that it actually uses the stdout from the spawned
subprocess. Up to this point panics would cause tests to fail, however, it did
not panic with the stdout of that subprocess. It would instead just print out
the panic message from the callsite in the macro. To rectify this we now install
a custom panic hook that will use the default if a special header is not passed
into a panic message. In the replaced hook though we print out the output of the
panic from the spawned subprocess now. With this we actually now show the proper
output to the caller, almost as if we never spawned a subprocess.

We now also have a way to test these by using one test to spawn broken tests
which use the assay macro, but are ignored otherwise. This way we can test bad
tests or failure cases and check that the output is what we expect!