Use v1 for dependencies of assay
In order to provide maximum compatibility for versioning for users and
reducing the number of deps in our users tree we set pretty_assertions_sorted,
tokio, and async-std to v1. We don't expect semver breaking changes with these
deps and so can let the user choose 1.x.y for their deps which should mean we
don't have two different versions compiled in our user's dep tree.