Add net module for testing network code
This is the first change to assay that is adding code to help write
tests not just making running them easier. This addition does two things

1. Creates a TestAddress trait that is implemented for TcpListener and
UdpSocket. This lets users bind TCP/UDP ports to an ipv4 or ipv6 address
with a random port that won't conflict as it let's the OS choose the
port

2. Imports this extention trait as part of the assay macro so that all a
user needs to do is import the TcpListener or UdpSocket type and calling
`(UdpSocket/TcpListener)::test_(v4/v6)()` will generate an address for
them

Assay will over time add more code like this to just make testing easier
for users.