Inline HOW_TO_USE into README/update licenses
This commit moves the HOW_TO_USE.md file inline to the README.md file.
The README file in it's entirety is now included in our rustdocs as
well. With the changes in the header sizes this makes it easy for users
to skip to the section whey want to learn how to use the macro for
themselves.

We also update the copyright and add it to all of the code files that
did not have them before this change.