Enhance error handling for improper macro usage
Up to this point using the macro was a source of frustration due to a
lack of good error messages if improperly used as a proc macro. With
this commit we now have helpful error messages and tests for various
scenarios! Users should be able to use assay with more confidence now.

Closes #13