Add ability to choose dest for included test files
This commit changes the proc macro so that you can now specify which
files to include using a tuple syntax of (src, dest) in the include
attribute of the assay macro. We also fixed an issue where multiple
subprocesses could spawn with simillar prefixes by passing the --exact
flag to cargo. Another issue was also fixed whereby passing --ignored
to cargo test would not actually run ignored tests and would allow the
code to falsely report that it was fine.

Closes #9