Add timeout feature to assay
Tests can now specify a timeout so that they fail after a certain period
of time. This is useful when you need something to complete in a short
amount of time and anything longer is a bug, even if the test would pass.
The README has been updated with how to use this new feature.